+++ example/example/settings.py
@@ -31,7 +31,14 @@ TEMPLATES = [
         "BACKEND": "django.template.backends.django.DjangoTemplates",
         "DIRS": [BASE_DIR / "templates"],
         "APP_DIRS": True,
-    }
+    },
+    {
+        "BACKEND": "django.template.backends.jinja2.Jinja2",
+        "DIRS": [BASE_DIR / "templates"],
+        "OPTIONS": {
+            "environment": "example.jinja.environment",
+        },
+    },
 ]
 
 USE_TZ = True