+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 * Mention django-browser-reload in reload debug message.
 
+* Add Jinja template tag.
+
 1.0.0 (2021-12-15)
 ------------------